refactor: 832 - mentorship folder
