+++ src/widgets/mentors-after-register/ui/mentors-after-register.module.scss
