+++ src/widgets/mentors-after-register/ui/mentors-after-register.tsx
