+++ src/widgets/mentorship-activities/ui/activity-card/activity-card.test.tsx
