+++ src/widgets/mentorship-activities/ui/mentor-activities/mentor-activities.module.scss
