+++ src/widgets/mentorship-courses/mentorship-courses.test.tsx
@@ -1,9 +1,9 @@
 import { screen, within } from '@testing-library/react';
 
-import { MentorshipCourseTitles } from '../../../../../dev-data/mentorship-data.types';
+import { MentorshipCourseTitles } from '../../../dev-data/mentorship-data.types';
 import { renderWithRouter } from '@/shared/__tests__/utils';
 import { mentorshipCourseTitles } from '@/views/mentorship/constants';
-import { MentorshipCourses } from '@/views/mentorship/ui/mentorship-courses/mentorship-courses';
+import { MentorshipCourses } from '@/widgets/mentorship-courses/mentorship-courses';
 
 describe('Mentorship courses', () => {
   it('renders the correct content', async () => {
@@ -20,7 +20,9 @@ describe('Mentorship courses', () => {
 
     courseCards.forEach((card) => {
       const courseTitle = within(card).getByTestId('subtitle');
-      const isMentorshipCourse = mentorshipCourseTitles.includes(courseTitle.textContent as MentorshipCourseTitles);
+      const isMentorshipCourse = mentorshipCourseTitles.includes(
+        courseTitle.textContent as MentorshipCourseTitles,
+      );
 
       expect(card).toBeVisible();
       expect(isMentorshipCourse).toBeTruthy();
