+++ src/widgets/mentorship-details/details/details.module.scss
