+++ src/widgets/mentorship-details/details/details.tsx
