+++ src/widgets/mentorship-hero/ui/mentorship-hero.module.scss
@@ -1,6 +1,6 @@
 .hero {
   min-height: 440px;
-  background: url('../../../../shared/assets/mentorship-hero.webp') center / cover no-repeat;
+  background: url('../../../shared/assets/mentorship-hero.webp') center / cover no-repeat;
 
   &-content {
     display: flex;
